+++ spec/controllers/followups_controller_spec.rb
@@ -32,6 +32,14 @@ describe FollowupsController do
         expect(response).to be_success
       end
 
+      it "displays 'wrong user' message when not logged in as the request owner" do
+        get :new, params: {
+                    request_id: request.id,
+                    incoming_message_id: message_id
+                  }
+        expect(response).to render_template('user/wrong_user')
+      end
+
       it 'raises an ActiveRecord::RecordNotFound error for other embargoed requests' do
         embargoed_request = FactoryBot.create(:embargoed_request)
         expect {
@@ -321,6 +329,26 @@ describe FollowupsController do
       session[:user_id] = request_user.id
     end
 
+    shared_examples_for 'successful_followup_sent' do
+
+      it 'sends the followup message' do
+        post :create, params: {
+                        outgoing_message: dummy_message,
+                        request_id: request.id,
+                        incoming_message_id: message_id
+                      }
+
+        # check it worked
+        deliveries = ActionMailer::Base.deliveries
+        expect(deliveries.size).to eq(1)
+        mail = deliveries[0]
+        expect(mail.body).to match(/What a useless response! You suck./)
+        expect(mail.to_addrs.first.to_s).
+          to eq(request.public_body.request_email)
+      end
+
+    end
+
     context "when not logged in" do
       before do
         session[:user_id] = nil
@@ -396,20 +424,7 @@ describe FollowupsController do
       expect(response).to render_template('new')
     end
 
-    it "sends the follow up message" do
-      post :create, params: {
-                      :outgoing_message => dummy_message,
-                      :request_id => request.id,
-                      :incoming_message_id => message_id
-                    }
-
-      # check it worked
-      deliveries = ActionMailer::Base.deliveries
-      expect(deliveries.size).to eq(1)
-      mail = deliveries[0]
-      expect(mail.body).to match(/What a useless response! You suck./)
-      expect(mail.to_addrs.first.to_s).to eq(request.public_body.request_email)
-    end
+    it_behaves_like 'successful_followup_sent'
 
     it "updates the status for successful followup sends" do
       post :create, params: {
@@ -419,6 +434,30 @@ describe FollowupsController do
                     }
 
       expect(request.reload.described_state).to eq('waiting_response')
+    end
+
+    context 'the request is no longer open to "anybody"' do
+
+      before do
+        request.update_attributes(
+          allow_new_responses_from: 'authority_only',
+          reject_incoming_at_mta: true
+        )
+      end
+
+      it 'reopens the parent request to new responses' do
+        post :create, params: {
+                        outgoing_message: dummy_message,
+                        request_id: request.id,
+                        incoming_message_id: message_id
+                      }
+
+        expect(request.reload.allow_new_responses_from).to eq('anybody')
+        expect(request.reload.reject_incoming_at_mta).to eq(false)
+      end
+
+      it_behaves_like 'successful_followup_sent'
+
     end
 
     it "updates the event status for successful followup sends if the request is waiting clarification" do